# Avoid allocating when comparing type names