--- a/src/Compiler/Microsoft.CodeAnalysis.Razor/src/SymbolExtensions.cs
+++ b/src/Compiler/Microsoft.CodeAnalysis.Razor/src/SymbolExtensions.cs
@@ -14,6 +14,23 @@ internal static class SymbolExtensions
 
     internal static bool HasFullName(this ISymbol symbol, string fullName)
     {
+        // generally we expect this call to fail, so try and match on the simple name first, before
+        // checking the fully qualified name, to prevent unnecesary string allocation from ToDisplayString 
+        if (symbol.Name.Length > fullName.Length)
+        {
+            return false;
+        }
+
+        // the simple name doesn't include generic type params, so skip over them in the full name if they are present
+        int fullNameEndIndex = fullName[^1] == '>' ? fullName.LastIndexOf('<') : fullName.Length;
+        int fullNameStartIndex = fullName.LastIndexOf('.', fullNameEndIndex - 1) + 1;
+
+        if (!fullName.AsSpan(fullNameStartIndex, fullNameEndIndex - fullNameStartIndex).Equals(symbol.Name.AsSpan(), StringComparison.Ordinal))
+        {
+            return false;
+        }
+
+        // The simple name matched, do a fully qualified comparison
         return symbol.ToDisplayString(FullNameTypeDisplayFormat).Equals(fullName, StringComparison.Ordinal);
     }
 }