Check the type simple name first, to avoid allocating a lot of strings via `ToDisplayString()`

In the cold benchmark we see a drop of about 6MB

```
Previous:
|           Method |     Mean |   Error |  StdDev |      Gen0 |      Gen1 | Allocated |
|----------------- |---------:|--------:|--------:|----------:|----------:|----------:|
| Cold_Compilation | 254.5 ms | 4.80 ms | 4.25 ms | 3000.0000 | 1000.0000 |  95.29 MB |

After:
|           Method |     Mean |   Error |  StdDev |      Gen0 |      Gen1 | Allocated |
|----------------- |---------:|--------:|--------:|----------:|----------:|----------:|
| Cold_Compilation | 251.6 ms | 2.77 ms | 2.31 ms | 3000.0000 | 1000.0000 |  89.25 MB |

```

We also see a drop of a 100kb or so on the hot benchmarks, but they already allocated very little.